+++ Audio_mart/src/main/java/com/audio_mart/mapper/MemberMapper.java
@@ -7,10 +7,10 @@ import com.audio_mart.domain.MemberDTO;
 @Mapper
 public interface MemberMapper {
 	public boolean insertMember(MemberDTO params);
-	public MemberDTO findByIdx(int idx);
+	public MemberDTO findByIdx(Long idx);
 	public MemberDTO findByCustid(String custid);
 	public boolean updateMember(MemberDTO params);
-	public boolean deleteMember(int idx);
+	public boolean deleteMember(Long idx);
 	public List<MemberDTO> memberList();
 //	public int selectMemberTotalCount();
 }
