+++ Audio_mart/src/main/java/com/audio_mart/service/ProductService.java
@@ -14,9 +14,9 @@ public interface ProductService {
 	// 상품 수정하기
 	public boolean updateProduct(ProductDTO params);
 	// 한 상품 정보 가져오기
-	public ProductDTO findByProductId(int ProductId);
+	public ProductDTO findByProductId(Long ProductId);
 	// 상품 삭제하기
-	public boolean removeProduct(int productId);
+	public boolean removeProduct(Long productId);
 	// 상품 목록 가져오기
 	public List<ProductDTO> getProductList();
